+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @clerk/testing
 
+## 1.4.23
+
+### Patch Changes
+
+- Fix captcha triggering on Cypress tests ([#5182](https://github.com/clerk/javascript/pull/5182)) by [@anagstef](https://github.com/anagstef)
+
+- Updated dependencies [[`d76c4699990b8477745c2584b1b98d5c92f9ace6`](https://github.com/clerk/javascript/commit/d76c4699990b8477745c2584b1b98d5c92f9ace6), [`a9b0087fca3f427f65907b358d9b5bc0c95921d8`](https://github.com/clerk/javascript/commit/a9b0087fca3f427f65907b358d9b5bc0c95921d8), [`92d17d7c087470b262fa5407cb6720fe6b17d333`](https://github.com/clerk/javascript/commit/92d17d7c087470b262fa5407cb6720fe6b17d333)]:
+  - @clerk/shared@2.22.0
+  - @clerk/types@4.46.1
+  - @clerk/backend@1.24.2
+
 ## 1.4.22
 
 ### Patch Changes